[SOIN] Extrais les types et fonctions pour évaluer les niveaux de risque
...afin de les réutiliser pour calculer le niveau d’un risque dans le tableau
