+++ svelte/lib/risques/risques.ts
@@ -1,10 +1,11 @@
 import { decode } from 'html-entities';
 import Risques from './Risques.svelte';
-import type {
-  DonneesRisque,
-  Risque,
-  RisquesProps,
-  TypeRisque,
+import {
+  type DonneesRisque,
+  NiveauRisque,
+  type Risque,
+  type RisquesProps,
+  type TypeRisque,
 } from './risques.d';
 
 document.body.addEventListener(
@@ -57,5 +58,36 @@ export const intituleRisque = (risque: Risque) =>
 
 export const risqueAMettreAJour = (risque: Risque) =>
   risque.type === 'SPECIFIQUE' && !risque.categories.length;
+
+const referentielNiveauRisque: NiveauRisque[][] = [
+  [
+    NiveauRisque.Moyen,
+    NiveauRisque.Moyen,
+    NiveauRisque.Eleve,
+    NiveauRisque.Eleve,
+  ],
+  [
+    NiveauRisque.Faible,
+    NiveauRisque.Moyen,
+    NiveauRisque.Eleve,
+    NiveauRisque.Eleve,
+  ],
+  [
+    NiveauRisque.Faible,
+    NiveauRisque.Faible,
+    NiveauRisque.Moyen,
+    NiveauRisque.Eleve,
+  ],
+  [
+    NiveauRisque.Faible,
+    NiveauRisque.Faible,
+    NiveauRisque.Moyen,
+    NiveauRisque.Moyen,
+  ],
+];
+
+export const niveauRisqueCellule = (colonne: number, ligne: number) => {
+  return referentielNiveauRisque[ligne][colonne];
+};
 
 export default app!;
